restructure directories for organization
