+++ data/helpers/users/users.js
@@ -1,4 +1,4 @@
-const db = require("../dbConfig.js");
+const db = require("../../dbConfig.js");
 
 module.exports = {
   registerUser,
